requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
sqlalchemy==1.4.53
databases[sqlite]==0.8.0
aiosqlite==0.19.0
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
schedule==1.2.0
loguru==0.7.2 